bugfix： 返回的数据结构字段变了。。 change：新增了签名文件，改了ignore，升级了kotlin，准备换图标

+++ app/src/main/java/com/sanousun/mdweather/model/AirQualityIndex.kt
@@ -16,7 +16,7 @@ data class AirQualityIndex(
             /**
              * 一氧化碳指数
              */
-            val co: Int,
+            val co: Float,
             /**
              * 一氧化氮指数
              */
